Add labelPairs.String() for logging
Print out labelPairs the same as Prometheus model.Metric.
Move labelPair code out to its own file.
